Fixed compiler warnings in tests
Former-commit-id: 59dadf338731d6fe308bc0fd5ad51bbfa5fd1a5c

+++ tests/ParticleTest.m
@@ -233,7 +233,7 @@ Class restartAction()
 {
 	CCLabelAtlas *atlas = (CCLabelAtlas*) [self getChildByTag:kTagParticleCount];
 
-	NSString *str = [NSString stringWithFormat:@"%4d", emitter_.particleCount];
+	NSString *str = [NSString stringWithFormat:@"%4ld", (unsigned long)emitter_.particleCount];
 	[atlas setString:str];
 }
